CU 작업 시작 4 DU 작업 시작 1
CU에는 SCTP 서버 추가
DU에는 SCTP 클라이언트 추가

+++ src/utils/common_types.hpp
@@ -174,13 +174,13 @@ enum class ECellCategory
 
 struct Vector3
 {
-    int x{};
-    int y{};
-    int z{};
+    double x{};
+    double y{};
+    double z{};
 
     Vector3() = default;
 
-    Vector3(int x, int y, int z) : x(x), y(y), z(z)
+    Vector3(double x, double y, double z) : x(x), y(y), z(z)
     {
     }
 };
